% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_join.R
\name{as_view}
\alias{as_view}
\title{as_view}
\usage{
as_view(x, tblname = tmp_tbl_name(), conn = cached_connection())
}
\arguments{
\item{x}{a duckdb spatial dataset}

\item{tblname}{The name of the table to create in the database.}

\item{conn}{A connection to a database.}
}
\description{
Create a View of the current query.  This can be an effective way to allow
a query chain to remain lazy
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
path <- system.file("extdata/spatial-test.csv", package="duckdbfs")
df <- open_dataset(path)
library(dplyr)

df |> filter(latitude > 5) |> as_view()
\dontshow{\}) # examplesIf}
}
