% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_connection.R
\name{close_connection}
\alias{close_connection}
\title{close connection}
\usage{
close_connection(conn = cached_connection())
}
\arguments{
\item{conn}{a duckdb connection (leave blank)
Closes the invisible cached connection to duckdb}
}
\value{
returns nothing.
}
\description{
close connection
}
\details{
Shuts down connection before gc removes it.
Then clear cached reference to avoid using a stale connection
This avoids complaint about connection being garbage collected.
}
\examples{

close_connection()

}
