% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_connection.R
\name{connect}
\alias{connect}
\alias{cached_connection}
\title{connect}
\usage{
connect(
  dbdir = ":memory:",
  read_only = FALSE,
  bigint = "numeric",
  config = list(temp_directory = tempfile()),
  autoload_exts = getOption("duckdbfs_autoload_extensions", TRUE),
  with_spatial = TRUE,
  with_h3 = TRUE
)
}
\description{
connect
}
