% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_config.R
\name{duckdb_get_config}
\alias{duckdb_get_config}
\title{duckdb reset configuration to default}
\usage{
duckdb_get_config(x = NULL, conn = cached_connection())
}
\arguments{
\item{x}{parameter name. Omit to see a table of all settings.}

\item{conn}{A connection to a database.}
}
\description{
duckdb reset configuration to default
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Full config table
duckdb_get_config()

# look up single config value
duckdb_get_config("threads")

# set a different value, test, reset.
duckdb_config(threads = 10)
duckdb_get_config("threads")
duckdb_reset("threads")
\dontshow{\}) # examplesIf}
}
\seealso{
duckdb_config, duckdb_get_config
}
