% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_config.R
\name{duckdb_reset}
\alias{duckdb_reset}
\title{duckdb reset configuration to default}
\usage{
duckdb_reset(x, conn = cached_connection())
}
\arguments{
\item{x}{parameter name}

\item{conn}{A connection to a database.}
}
\description{
duckdb reset configuration to default
}
\examples{
duckdb_config(threads = 10)
duckdb_get_config("threads")
duckdb_reset("threads")
}
\seealso{
duckdb_config, duckdb_get_config
}
