% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_config.R
\name{load_spatial}
\alias{load_spatial}
\title{load the duckdb geospatial data plugin}
\usage{
load_spatial(conn = cached_connection())
}
\arguments{
\item{conn}{A database connection object created using the
\code{cache_connection} function (default: \code{cache_connection()}).}
}
\value{
loads the extension and returns status invisibly.
}
\description{
load the duckdb geospatial data plugin
}
\references{
\url{https://duckdb.org/docs/extensions/spatial.html}
}
