% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_geojson.R
\name{to_geojson}
\alias{to_geojson}
\title{Write geojson using duckdb's native JSON writer}
\usage{
to_geojson(
  dataset,
  path,
  conn = cached_connection(),
  id_col = NULL,
  as_http = FALSE,
  server = Sys.getenv("AWS_S3_ENDPOINT", "s3.amanzonaws.com"),
  use_ssl = Sys.getenv("AWS_HTTPS", "TRUE")
)
}
\arguments{
\item{dataset}{a remote tbl object from \code{open_dataset},
or an in-memory data.frame.}

\item{path}{a local file path or S3 path with write credentials}

\item{conn}{duckdbfs database connection}

\item{id_col}{(deprecated). to_geojson() will preserve all atomic columns
as properties.}

\item{as_http}{convert returned S3 path to URL (e.g. for public buckets)}

\item{server}{aws endpoint if converting s3 path to URL}

\item{use_ssl}{should url use https}
}
\value{
path, invisibly
}
\description{
Write geojson using duckdb's native JSON writer
}
