% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3.R
\name{to_h3j}
\alias{to_h3j}
\title{Write H3 hexagon data out as an h3j-compliant JSON file
NOTE: the column containing H3 hashes must be named \code{hexid}}
\usage{
to_h3j(x, dest, con = cached_connection())
}
\description{
Write H3 hexagon data out as an h3j-compliant JSON file
NOTE: the column containing H3 hashes must be named \code{hexid}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# example code
\dontshow{\}) # examplesIf}
}
