% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_json.R
\name{to_json}
\alias{to_json}
\title{to_json
write data out as a JSON object}
\usage{
to_json(x, dest, con = cached_connection(), array = TRUE, options)
}
\arguments{
\item{options}{additional options as a char string, see}
}
\description{
to_json
write data out as a JSON object
}
