% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataset.R
\name{write_dataset}
\alias{write_dataset}
\title{write_dataset}
\usage{
write_dataset(dataset, path, conn = cached_connection(), ...)
}
\arguments{
\item{dataset}{a remote tbl object from \code{open_dataset},
or an in-memory data.frame.}

\item{path}{a local file path or S3 path with write credentials}

\item{conn}{duckdbfs database connection}

\item{...}{additional arguments to \code{\link[=duckdb_s3_config]{duckdb_s3_config()}}}
}
\description{
write_dataset
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
  write_dataset(mtcars, tempfile())
\dontshow{\}) # examplesIf}
}
