% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataset.R
\name{write_geo}
\alias{write_geo}
\title{Write a geojson file}
\usage{
write_geo(dataset, path, conn = cached_connection())
}
\arguments{
\item{dataset}{a remote tbl object from \code{open_dataset},
or an in-memory data.frame.}

\item{path}{a local file path or S3 path with write credentials}

\item{conn}{duckdbfs database connection}
}
\description{
Write a geojson file
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
local_file <-  system.file("extdata/spatial-test.csv", package="duckdbfs")
load_spatial()
tbl <- open_dataset(local_file, format='csv')
write_geo(tbl, "spatial.geojson")
\dontshow{\}) # examplesIf}
}
