% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_config.R
\name{duckdb_config}
\alias{duckdb_config}
\title{duckdb configuration}
\usage{
duckdb_config(..., conn = cached_connection())
}
\arguments{
\item{...}{named argument of the parameters to set, see examples
see all possible configuration options at \url{https://duckdb.org/docs/sql/configuration.html}}

\item{conn}{A connection to a database.}
}
\value{
the active duckdb connection, invisibly
}
\description{
duckdb configuration
}
\details{
Note: in I/O bound tasks such as streaming data, it can be helpful to set
thread parallelism signifantly higher than avialable CPU cores.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
duckdb_config(threads = 1, memory_limit = '10GB')
duckdb_get_config("threads")
duckdb_reset("threads")
\dontshow{\}) # examplesIf}
}
\seealso{
duckdb_reset, duckdb_get_config
}
