% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_config.R
\name{load_spatial}
\alias{load_spatial}
\title{load the duckdb geospatial data plugin}
\usage{
load_spatial(
  conn = cached_connection(),
  nightly = getOption("duckdbfs_use_nightly", FALSE),
  force = FALSE
)
}
\arguments{
\item{conn}{A database connection object created using the
\code{cache_connection} function (default: \code{cache_connection()}).}

\item{nightly}{should we use the nightly version or not?
default FALSE, configurable as \code{duckdbfs_use_nightly} option.}

\item{force}{force re-install?}
}
\value{
loads the extension and returns status invisibly.
}
\description{
load the duckdb geospatial data plugin
}
\references{
\url{https://duckdb.org/docs/extensions/spatial.html}
}
