% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_h3j.R
\name{to_h3j}
\alias{to_h3j}
\title{Write H3 hexagon data out as an h3j-compliant JSON file
NOTE: the column containing H3 hashes must be named \code{hexid}}
\usage{
to_h3j(dataset, path, conn = cached_connection(), as_http = FALSE)
}
\arguments{
\item{dataset}{a remote tbl object from \code{open_dataset},
or an in-memory data.frame.}

\item{path}{a local file path or S3 path with write credentials}

\item{conn}{duckdbfs database connection}

\item{as_http}{if path is an S3 location, will return corresponding HTTP address.}
}
\description{
Write H3 hexagon data out as an h3j-compliant JSON file
NOTE: the column containing H3 hashes must be named \code{hexid}
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# example code
\dontshow{\}) # examplesIf}
}
