% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_json.R
\name{to_json}
\alias{to_json}
\title{to_json
write data out as a JSON object}
\usage{
to_json(
  dataset,
  path,
  conn = cached_connection(),
  array = TRUE,
  options = NULL
)
}
\arguments{
\item{dataset}{a remote tbl object from \code{open_dataset},
or an in-memory data.frame.}

\item{path}{a local file path or S3 path with write credentials}

\item{conn}{duckdbfs database connection}

\item{array}{generate a JSON array?}

\item{options}{additional options as a char string, see}
}
\description{
to_json
write data out as a JSON object
}
